+++ changelog.md
@@ -1,5 +1,16 @@
 # Change Log for WTToolBox
 
+## 1.1.0
+
++ Added `Test-WTVersion`
++ Modified module to create a global variable `$WTSettings` with a converted version of `settings.json`.
++ Modified module to create a global variable `WTSettingsPath` with the path to the `setting.json` file.
++ Modified module to create a global variable `$WTDefaults` with a converted version of `defaults.json`.
++ Added a private function called `AddWTSettingsVariable    .
++ Updated `README.md`.
++ Minor help revisions
++ Updated Pester test file
+
 ## 1.0.0
 
 + Updated module manifest